chore(test): test cases for bugs
