+++ package.json
@@ -67,6 +67,8 @@
     "prepublishOnly": "npm run dist",
     "screenshot": "node ./bin/screenshot.js",
     "test": "torch --compile --renderer --recursive ./test/unit",
+    "test-bugs": "torch --compile --renderer --recursive ./test/bugs",
+    "test-bugs-live": "torch --compile --interactive --watch --recursive ./test/bugs",
     "test-live": "torch --compile --interactive --watch --recursive ./test/unit",
     "watch": "webpack --config webpack-dev.config.js",
     "win-dev": "node ./bin/win-dev.js"
@@ -74,7 +76,8 @@
   "pre-commit": {
     "run": [
       "lint",
-      "test"
+      "test",
+      "test-bugs"
     ],
     "silent": false
   },
